Refactor: create `Stroke::round_center_to_pixel`
